// src/views/personal/keyword/zhuanfa/utils/types.ts
// 虽然字段很少 但是抽离出来 后续有扩展字段需求就很方便了

interface FormItemProps {
  id?: number;
  /** 用于判断是`新增`还是`修改` */
  title: string;
  /** 关键词名称 */
  keyword: string;
  /** 回复内容 */
  answer: string;
  /** 关键词匹配信息 */
  data?: any;
  user: any[];
  group: any[];
  /** 关键词状态 */
  enable?: string;
}
interface FormProps {
  formInline: FormItemProps;
}

export type { FormItemProps, FormProps };